refactor: set to uint104

+++ solidity/contracts/adapters/UniswapV3Adapter.sol
@@ -26,7 +26,7 @@ contract UniswapV3Adapter is AccessControl, SimpleOracle, IUniswapV3Adapter {
   /// @inheritdoc IUniswapV3Adapter
   uint8 public cardinalityPerMinute;
   /// @inheritdoc IUniswapV3Adapter
-  uint216 public gasPerCardinality = 22_250;
+  uint104 public gasPerCardinality = 22_250;
 
   mapping(bytes32 => bool) internal _isPairDenylisted; // key(tokenA, tokenB) => is denylisted
   mapping(bytes32 => address[]) internal _poolsForPair; // key(tokenA, tokenB) => pools
@@ -110,7 +110,7 @@ contract UniswapV3Adapter is AccessControl, SimpleOracle, IUniswapV3Adapter {
   }
 
   /// @inheritdoc IUniswapV3Adapter
-  function setGasPerCardinality(uint216 _gasPerCardinality) external onlyRole(ADMIN_ROLE) {
+  function setGasPerCardinality(uint104 _gasPerCardinality) external onlyRole(ADMIN_ROLE) {
     if (_gasPerCardinality == 0) revert InvalidGasPerCardinality();
     gasPerCardinality = _gasPerCardinality;
     emit GasPerCardinalityChanged(_gasPerCardinality);
@@ -151,7 +151,7 @@ contract UniswapV3Adapter is AccessControl, SimpleOracle, IUniswapV3Adapter {
     // Load to mem to avoid multiple storage reads
     address[] storage _storagePools = _poolsForPair[_pairKey];
     uint256 _poolsPreviouslyInStorage = _storagePools.length;
-    uint216 _gasCostPerCardinality = gasPerCardinality;
+    uint104 _gasCostPerCardinality = gasPerCardinality;
 
     uint16 _targetCardinality = uint16((period * cardinalityPerMinute) / 60) + 1;
     uint256 _preparedPools;
@@ -159,7 +159,7 @@ contract UniswapV3Adapter is AccessControl, SimpleOracle, IUniswapV3Adapter {
       address _pool = _pools[i];
       (, , , , uint16 _currentCardinality, , ) = IUniswapV3Pool(_pool).slot0();
       if (_currentCardinality < _targetCardinality) {
-        uint216 _gasCostToIncreaseAndAddSupport = uint216(_targetCardinality - _currentCardinality) *
+        uint104 _gasCostToIncreaseAndAddSupport = uint104(_targetCardinality - _currentCardinality) *
           _gasCostPerCardinality +
           _FIXED_GAS_COST_TO_SUPPORT_POOL;
         if (_gasCostToIncreaseAndAddSupport > gasleft()) {
